+++ packages/permissionless/accounts/safe/toSafeSmartAccount.ts
@@ -1348,9 +1348,15 @@ export async function toSafeSmartAccount<
                 )
                 operationType = 1
             } else {
-                to = calls[0].to
-                data = calls[0].data ?? "0x"
-                value = calls[0].value ?? 0n
+                const call = calls.length === 0 ? undefined : calls[0]
+
+                if (!call) {
+                    throw new Error("No calls to encode")
+                }
+
+                to = call.to
+                data = call.data ?? "0x"
+                value = call.value ?? 0n
             }
 
             return encodeFunctionData({
@@ -1363,7 +1369,7 @@ export async function toSafeSmartAccount<
             return getAccountNonce(client, {
                 address: await this.getAddress(),
                 entryPointAddress: entryPoint.address,
-                key: args?.key ?? nonceKey
+                key: nonceKey ?? args?.key
             })
         },
         async getStubSignature() {