+++ packages/permissionless/accounts/simple/toSimpleSmartAccount.ts
@@ -256,6 +256,12 @@ export async function toSimpleSmartAccount<
                 })
             }
 
+            const call = calls.length === 0 ? undefined : calls[0]
+
+            if (!call) {
+                throw new Error("No calls to encode")
+            }
+
             return encodeFunctionData({
                 abi: [
                     {
@@ -283,14 +289,14 @@ export async function toSimpleSmartAccount<
                     }
                 ],
                 functionName: "execute",
-                args: [calls[0].to, calls[0].value ?? 0n, calls[0].data ?? "0x"]
+                args: [call.to, call.value ?? 0n, call.data ?? "0x"]
             })
         },
         async getNonce(args) {
             return getAccountNonce(client, {
                 address: await this.getAddress(),
                 entryPointAddress: entryPoint.address,
-                key: args?.key ?? nonceKey
+                key: nonceKey ?? args?.key
             })
         },
         async getStubSignature() {