+++ packages/permissionless/actions/pimlico/getTokenQuotes.ts
@@ -24,6 +24,7 @@ export type GetTokenQuotesReturnType = {
     token: Address
     postOpGas: bigint
     exchangeRate: bigint
+    exchangeRateNativeToUsd: bigint
 }[]
 
 /**
@@ -62,6 +63,7 @@ export const getTokenQuotes = async <
     return res.quotes.map((quote) => ({
         ...quote,
         postOpGas: hexToBigInt(quote.postOpGas),
-        exchangeRate: hexToBigInt(quote.exchangeRate)
+        exchangeRate: hexToBigInt(quote.exchangeRate),
+        exchangeRateNativeToUsd: hexToBigInt(quote.exchangeRateNativeToUsd)
     }))
 }