+++ packages/permissionless/types/pimlico.ts
@@ -34,6 +34,7 @@ type GetTokenQuotesWithBigIntAsHex = {
         token: Address
         postOpGas: Hex
         exchangeRate: Hex
+        exchangeRateNativeToUsd: Hex
     }[]
 }