+++ packages/wagmi/index.ts
@@ -1,32 +1,7 @@
-import {
-    type SimpleSmartAccountParameters,
-    simpleSmartAccount
-} from "./connectors/simpleSmartAccount"
-import { smartAccount } from "./connectors/smartAccount"
-
-import {
-    type SafeSmartAccountParameters,
-    safeSmartAccount
-} from "./connectors/safeSmartAccount"
-
-import {
-    type BiconomySmartAccountParameters,
-    biconomySmartAccount
-} from "./connectors/biconomySmartAccount"
-
-import {
-    type KernelSmartAccountParameters,
-    kernelSmartAccount
-} from "./connectors/kernelSmartAccount"
-
 export {
-    smartAccount,
-    simpleSmartAccount,
-    type SimpleSmartAccountParameters,
-    safeSmartAccount,
-    type SafeSmartAccountParameters,
-    biconomySmartAccount,
-    type BiconomySmartAccountParameters,
-    kernelSmartAccount,
-    type KernelSmartAccountParameters
-}
+    PermissionlessProvider,
+    type PermissionlessProviderProps
+} from "./context"
+
+export { useSendTransaction } from "./hooks/useSendTransaction"
+export { useWaitForTransactionReceipt } from "./hooks/useWaitForTransactionReceipt"